@@ -1,6 +1,6 @@
 module allhash
 
-go 1.22
+go 1.25
 
 require (
 	github.com/glaslos/ssdeep v0.4.0
