@@ -1,13 +1,10 @@
 module github.com/mosajjal/go-exp/elasticdump
 
-go 1.22
+go 1.25
 
-require github.com/sirupsen/logrus v1.9.3
-
-require github.com/tidwall/match v1.1.1 // indirect
+require github.com/tidwall/gjson v1.18.0
 
 require (
-	github.com/tidwall/gjson v1.17.1
+	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.1 // indirect
-	golang.org/x/sys v0.21.0 // indirect
 )
