@@ -1,6 +1,6 @@
 module github.com/mosajjal/go-exp/gopherchain
 
-go 1.23.4
+go 1.24.4
 
 require (
 	github.com/folbricht/routedns v0.1.88
