@@ -1,3 +1,3 @@
 module github.com/mosajjal/go-exp/identme
 
-go 1.22
+go 1.25
