@@ -1,3 +1,3 @@
 module github.com/mosajjal/go-exp/mhs
 
-go 1.23.3
+go 1.25.3
