@@ -1,3 +1,3 @@
 module github.com/mosajjal/go-exp/oauth2
 
-go 1.23.2
+go 1.25.2
