@@ -1,6 +1,6 @@
 module proxyenv
 
-go 1.22.0
+go 1.25.0
 
 require (
 	github.com/elazarl/goproxy v0.0.0-20231117061959-7cc037d33fb5
