@@ -1,5 +1,6 @@
-// fakedns is a simple DNS provider that returns a set of static IPs for given domain
-// prefix, suffix and FQDNs. It can fall back to an upstream DNS for the unmatched domains.
+// Package main provides fakedns, a programmable DNS server that returns static IPs
+// for configured domain patterns. It supports prefix, suffix, and FQDN matching with
+// upstream DNS fallback for unmatched domains.
 package main
 
 import (
@@ -7,6 +8,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"log/slog"
 	"net"
 	"net/http"
 	"os"
@@ -15,39 +17,45 @@ import (
 	"time"
 
 	"github.com/golang-collections/collections/tst"
-	"github.com/mosajjal/dnsclient"
-	slog "golang.org/x/exp/slog"
-
 	"github.com/miekg/dns"
+	"github.com/mosajjal/dnsclient"
 )
 
+// FakeDNS represents the DNS server configuration and routing tables.
 type FakeDNS struct {
 	dnsclient.Client
-	routePrefixes *tst.TernarySearchTree
-	routeSuffixes *tst.TernarySearchTree
-	routeFQDNs    map[string]TreeValue
+	routePrefixes *tst.TernarySearchTree // TST for prefix matching
+	routeSuffixes *tst.TernarySearchTree // TST for suffix matching (reversed strings)
+	routeFQDNs    map[string]TreeValue   // Map for exact FQDN matching
 	UDPPort       uint64
 	TCPPort       uint64
 }
 
-var fakeDNS = FakeDNS{}
-
 var (
-	matchPrefix = uint8(1)
-	matchSuffix = uint8(2)
-	matchFQDN   = uint8(3)
+	fakeDNS FakeDNS
+	dnsLock sync.RWMutex
+	logger  *slog.Logger
 )
 
-// Treevalue is inserted into TSTs as value for each prefix, suffix and FQDN
+const (
+	// Match types for domain routing
+	matchPrefix uint8 = 1
+	matchSuffix uint8 = 2
+	matchFQDN   uint8 = 3
+)
+
+// TreeValue is inserted into TSTs as value for each prefix, suffix and FQDN.
 type TreeValue struct {
-	Entry string
-	Mode  uint8
-	IP    net.IP
+	Entry string  // The original domain pattern
+	Mode  uint8   // Match type (prefix, suffix, or FQDN)
+	IP    net.IP  // IP address to return for this pattern
 }
 
-var dnsLock sync.RWMutex
-var log = slog.New(slog.NewTextHandler(os.Stderr, nil))
-var dnslog = slog.New(log.Handler().WithAttrs([]slog.Attr{{Key: "service", Value: slog.StringValue("dns")}}))
+func init() {
+	logger = slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{
+		Level: slog.LevelInfo,
+	}))
+}
 
 // inDomainList returns true if the domain is meant to be SKIPPED and not go through sni proxy
 func (c FakeDNS) inDomainList(fqdn string) bool {
@@ -114,10 +122,10 @@ func (c *FakeDNS) LoadDomainsCsv(Filename string) error {
 	c.routeSuffixes = tst.New()
 	c.routeFQDNs = make(map[string]TreeValue)
 
-	dnslog.Info("Loading the domain from file/url")
+	logger.Info("Loading the domain from file/url")
 	var scanner *bufio.Scanner
 	if strings.HasPrefix(Filename, "http://") || strings.HasPrefix(Filename, "https://") {
-		dnslog.Info("domain list is a URL, trying to fetch")
+		logger.Info("domain list is a URL, trying to fetch")
 		client := http.Client{
 			CheckRedirect: func(r *http.Request, via []*http.Request) error {
 				r.URL.Opaque = r.URL.Path
@@ -126,10 +134,10 @@ func (c *FakeDNS) LoadDomainsCsv(Filename string) error {
 		}
 		resp, err := client.Get(Filename)
 		if err != nil {
-			dnslog.Error("", err)
+			logger.Error("", err)
 			return err
 		}
-		dnslog.Info("(re)fetching URL", "url", Filename)
+		logger.Info("(re)fetching URL", "url", Filename)
 		defer resp.Body.Close()
 		scanner = bufio.NewScanner(resp.Body)
 
@@ -138,17 +146,17 @@ func (c *FakeDNS) LoadDomainsCsv(Filename string) error {
 		if err != nil {
 			return err
 		}
-		dnslog.Info("(re)loading File", "file", Filename)
+		logger.Info("(re)loading File", "file", Filename)
 		defer file.Close()
 		scanner = bufio.NewScanner(file)
 	}
 
 	for scanner.Scan() {
 		lowerCaseLine := strings.ToLower(scanner.Text())
-		// split the line by comma to understand thednslog.c
+		// split the line by comma to understand thelogger.c
 		fqdn := strings.Split(lowerCaseLine, ",")
 		if len(fqdn) != 3 {
-			dnslog.Info(lowerCaseLine + " is not a valid line, assuming FQDN")
+			logger.Info(lowerCaseLine + " is not a valid line, assuming FQDN")
 			fqdn = []string{lowerCaseLine, "fqdn"}
 		}
 		// add the fqdn to the hashtable with its type
@@ -163,12 +171,12 @@ func (c *FakeDNS) LoadDomainsCsv(Filename string) error {
 		case "fqdn":
 			c.routeFQDNs[fqdn[0]] = TreeValue{Mode: matchFQDN, IP: net.ParseIP(fqdn[2]), Entry: fqdn[0]}
 		default:
-			//dnslog.Warnf("%s is not a valid line, assuming fqdn", lowerCaseLine)
-			dnslog.Info(lowerCaseLine + " is not a valid line, assuming FQDN")
+			//logger.Warnf("%s is not a valid line, assuming fqdn", lowerCaseLine)
+			logger.Info(lowerCaseLine + " is not a valid line, assuming FQDN")
 			c.routeFQDNs[fqdn[0]] = TreeValue{Mode: matchFQDN, IP: net.ParseIP(fqdn[2]), Entry: fqdn[0]}
 		}
 	}
-	dnslog.Info(fmt.Sprintf("%s loaded with %d prefix, %d suffix and %d fqdn", Filename, c.routePrefixes.Len(), c.routeSuffixes.Len(), len(c.routeFQDNs)-c.routePrefixes.Len()-c.routeSuffixes.Len()))
+	logger.Info(fmt.Sprintf("%s loaded with %d prefix, %d suffix and %d fqdn", Filename, c.routePrefixes.Len(), c.routeSuffixes.Len(), len(c.routeFQDNs)-c.routePrefixes.Len()-c.routeSuffixes.Len()))
 
 	return nil
 }
@@ -189,7 +197,7 @@ func (c *FakeDNS) performExternalAQuery(fqdn string) ([]dns.RR, time.Duration, e
 	res, trr, err := c.Client.Query(context.Background(), &msg)
 	if err != nil {
 		if err.Error() == "EOF" {
-			dnslog.Info("reconnecting DNS...")
+			logger.Info("reconnecting DNS...")
 			// dnsc.C.Close()
 			// dnsc.C, err = dnsclient.New(c.UpstreamDNS, true)
 			err = c.Client.Reconnect()
@@ -207,7 +215,7 @@ func (c FakeDNS) processQuestion(q dns.Question) ([]dns.RR, error) {
 			return nil, err
 		}
 
-		dnslog.Info("returned sniproxy address for domain", "fqdn", q.Name)
+		logger.Info("returned sniproxy address for domain", "fqdn", q.Name)
 
 		return []dns.RR{rr}, nil
 	}
@@ -218,7 +226,7 @@ func (c FakeDNS) processQuestion(q dns.Question) ([]dns.RR, error) {
 		return nil, err
 	}
 
-	dnslog.Info("[DNS] returned origin address", "fqdn", q.Name, "rtt", rtt)
+	logger.Info("[DNS] returned origin address", "fqdn", q.Name, "rtt", rtt)
 
 	return resp, nil
 }
@@ -258,7 +266,7 @@ func handleDNS(w dns.ResponseWriter, r *dns.Msg) {
 	for _, q := range m.Question {
 		answers, err := fakeDNS.processQuestion(q)
 		if err != nil {
-			dnslog.Error("", err)
+			logger.Error("", err)
 			continue
 		}
 		m.Answer = append(m.Answer, answers...)
@@ -283,7 +291,7 @@ func main() {
 	// set up upstream DNS
 	dnsClient, err := dnsclient.New(*upstreamDNS, true, "")
 	if err != nil {
-		log.Error("Failed to create DNS client", err)
+		logger.Error("Failed to create DNS client", err)
 		panic(1)
 	}
 	fakeDNS.Client = dnsClient
@@ -292,19 +300,19 @@ func main() {
 	if *ruleFile != "" {
 		err = fakeDNS.LoadDomainsCsv(*ruleFile)
 		if err != nil {
-			log.Error("Failed to load rule", err)
+			logger.Error("Failed to load rule", err)
 			panic(1)
 		}
 	}
 	if fakeDNS.UDPPort != 0 {
 		go func() {
 			serverUDP := &dns.Server{Addr: fmt.Sprintf(":%d", fakeDNS.UDPPort), Net: "udp"}
-			dnslog.Info("Started UDP DNS", "host", "0.0.0.0", "port", fakeDNS.UDPPort)
+			logger.Info("Started UDP DNS", "host", "0.0.0.0", "port", fakeDNS.UDPPort)
 			err := serverUDP.ListenAndServe()
 			defer serverUDP.Shutdown()
 			if err != nil {
-				dnslog.Error("Error starting UDP DNS server", err)
-				dnslog.Info(fmt.Sprintf("Failed to start server: %s\nYou can run the following command to pinpoint which process is listening on port %d\nsudo ss -pltun -at '( dport = :%d or sport = :%d )'", err.Error(), fakeDNS.TCPPort, fakeDNS.TCPPort, fakeDNS.TCPPort))
+				logger.Error("Error starting UDP DNS server", err)
+				logger.Info(fmt.Sprintf("Failed to start server: %s\nYou can run the following command to pinpoint which process is listening on port %d\nsudo ss -pltun -at '( dport = :%d or sport = :%d )'", err.Error(), fakeDNS.TCPPort, fakeDNS.TCPPort, fakeDNS.TCPPort))
 				panic(2)
 			}
 		}()
@@ -313,12 +321,12 @@ func main() {
 	if fakeDNS.TCPPort != 0 {
 		go func() {
 			serverTCP := &dns.Server{Addr: fmt.Sprintf(":%d", fakeDNS.TCPPort), Net: "tcp"}
-			dnslog.Info("Started TCP DNS", "host", "0.0.0.0", "port", fakeDNS.TCPPort)
+			logger.Info("Started TCP DNS", "host", "0.0.0.0", "port", fakeDNS.TCPPort)
 			err := serverTCP.ListenAndServe()
 			defer serverTCP.Shutdown()
 			if err != nil {
-				dnslog.Error("Failed to start server", err)
-				dnslog.Info(fmt.Sprintf("You can run the following command to pinpoint which process is listening on port %d\nsudo ss -pltun -at '( dport = :%d or sport = :%d )'", fakeDNS.TCPPort, fakeDNS.TCPPort, fakeDNS.TCPPort))
+				logger.Error("Failed to start server", err)
+				logger.Info(fmt.Sprintf("You can run the following command to pinpoint which process is listening on port %d\nsudo ss -pltun -at '( dport = :%d or sport = :%d )'", fakeDNS.TCPPort, fakeDNS.TCPPort, fakeDNS.TCPPort))
 			}
 		}()
 	}
@@ -328,7 +336,7 @@ func main() {
 	// 	go func() {
 	// 		crt, err := tls.LoadX509KeyPair(c.TLSCert, c.TLSKey)
 	// 		if err != nil {
-	// 			dnslog.Error("", err)
+	// 			logger.Error("", err)
 	// 			panic(2)
 
 	// 		}
@@ -336,11 +344,11 @@ func main() {
 	// 		tlsConfig.Certificates = []tls.Certificate{crt}
 
 	// 		serverTLS := &dns.Server{Addr: ":853", Net: "tcp-tls", TLSConfig: tlsConfig}
-	// 		dnslog.Info("Started DoT DNS", "host", "0.0.0.0", "port", 853)
+	// 		logger.Info("Started DoT DNS", "host", "0.0.0.0", "port", 853)
 	// 		err = serverTLS.ListenAndServe()
 	// 		defer serverTLS.Shutdown()
 	// 		if err != nil {
-	// 			dnslog.Error("", err)
+	// 			logger.Error("", err)
 	// 		}
 	// 	}()
 	// }
@@ -349,7 +357,7 @@ func main() {
 
 	// 	crt, err := tls.LoadX509KeyPair(c.TLSCert, c.TLSKey)
 	// 	if err != nil {
-	// 		dnslog.Error("", err)
+	// 		logger.Error("", err)
 	// 	}
 	// 	tlsConfig := &tls.Config{}
 	// 	tlsConfig.Certificates = []tls.Certificate{crt}
@@ -357,11 +365,11 @@ func main() {
 	// 	// Create the QUIC listener
 	// 	doqServer, err := doqserver.New(":8853", crt, "127.0.0.1:53", true)
 	// 	if err != nil {
-	// 		dnslog.Error("", err)
+	// 		logger.Error("", err)
 	// 	}
 
 	// 	// Accept QUIC connections
-	// 	dnslog.Info("Starting QUIC listener on :8853")
+	// 	logger.Info("Starting QUIC listener on :8853")
 	// 	go doqServer.Listen()
 
 	// }
